Fix trackErrors "hint" data
The hint data is incorrectly implemented and causing major issues:

1. `event_id` is a hex value and any other value cause the error to not
   be reported.
2. `data` in the hint is only used locally before sending the event to
   Sentry.

Instead we will use the `scope` API to add the proper event tagging and
metadata. We can leverage the `addMetadataToContext` function to
consistently process metadata for a Sentry event.
